+++ backend/app/schemas/zone.py
@@ -8,7 +8,7 @@ class Zone(BaseModel):
     video_name: str
     zone_name: str
     zone_type: Literal["RESTRICTED", "WALKWAY"]
-    dwell_threshold_seconds: int = 0
+    dwell_threshold_seconds: float = 0.0
     is_active: bool = True
     ui_shape_data: str  # JSON string
     flattened_coordinates: str  # JSON string
